+++ app/helpers/treatment_arms_helper.rb
@@ -1,0 +1,2 @@
+module TreatmentArmsHelper
+end