+++ app/models/randomization_scheme.rb
@@ -8,10 +8,12 @@ class RandomizationScheme < ActiveRecord::Base
   # Model Validation
   validates_presence_of :name, :user_id, :project_id
   validates_uniqueness_of :name, scope: [:deleted, :project_id]
+  validates_numericality_of :randomization_goal, greater_than_or_equal_to: 0, only_integer: true
 
   # Model Relationships
   belongs_to :user
   belongs_to :project
+  has_many :treatment_arms, -> { where deleted: false }
 
   # Model Methods