+++ app/helpers/block_size_multipliers_helper.rb
@@ -1,0 +1,2 @@
+module BlockSizeMultipliersHelper
+end