+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Project owners and editors can now add one or more randomization schemes to a project
   - **Permuted-block Configuration**
     - Treatment Arms can now be specified along with a weight allocation
+    - Different block sizes and allocations can now be specified
 - **Gem Changes**
   - Updated to rails 4.2.3
   - Updated to contour 3.0.1.rc