+++ CHANGELOG.md
@@ -9,7 +9,7 @@
     - Stratification Factors with options can now be specified
 - **Gem Changes**
   - Updated to rails 4.2.3
-  - Updated to contour 3.0.1.rc
+  - Updated to contour 3.0.1
   - Updated to redcarpet 3.3.2
 
 ## 0.28.1  (June 10, 2015)